requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
django-rest-swagger
django-rest-authtoken